Create separate helper funtions
Functions for creating:
- test keys on the fly
- test tokens on the fly
- moved from userauth_test
